fix id-> uuid

+++ src/app/controllers/aseessment.controller.ts
@@ -2,6 +2,7 @@ import { NextFunction, Response, Request } from 'express';
 import AssessmentRepository from '@repositories/assessment.repository';
 import Assessment_game_typeRepository from '@repositories/assessment_game_type.repository';
 import TestRepository from '@repositories/test.repository';
+import Candidates_assessmentsRepository from '@repositories/candidates_assessments.repository';
 import CandidateRepository from '@repositories/candidate.repository';
 import Hr_game_typeRepository from '@repositories/hr_game_type.repository';
 import { BaseController } from './base.controller';
@@ -9,9 +10,9 @@ import { Authorized, UseBefore, BadRequestError, CurrentUser, Body, Get, JsonCon
 import { AuthMiddleware } from '@middlewares/auth.middleware';
 import { Service } from 'typedi';
 import {AsssessmentDto} from '../../dtos/assessment.dto'
-import { AuthRequest } from '@interfaces/response.interface';
-import { TestDto } from 'dtos/test.dto';
-import Candidate from '@models/entities/candidates';
+import { AuthRequest } from '@interfaces/response.interface';;
+import nodemailer from "nodemailer";
+import { GameType } from '@enum/game.enum';
 
 
 
@@ -25,6 +26,7 @@ class AssessmentController extends BaseController {
     protected testRepository: TestRepository,
     protected hr_game_typeRepository: Hr_game_typeRepository,
     protected candidateRepository: CandidateRepository,
+    protected candidates_assessmentsRepository: Candidates_assessmentsRepository,
   )
   {
     super();
@@ -36,35 +38,37 @@ class AssessmentController extends BaseController {
   async create(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
     console.log('a')
     try {
-      
+        let isSuccess = 1
         const assessDto: AsssessmentDto = req.body;
         const hr = req.hr
-        const {name, logical, memory, start_date, end_date} = assessDto;
+        const {name, game_types, start_date, end_date} = assessDto;
         //check if hr hr has right to approach the game type
-        if (logical)
-        console.log(hr.id)
-        var hasLogicalType = await this.hr_game_typeRepository.findByCondition({where:{hr_id: hr.id, game_type_id: 1}})
-        if (memory)
-        var hasMemoryType = await this.hr_game_typeRepository.findByCondition({where:{hr_id: hr.id, game_type_id: 2}})
-        if (!hasLogicalType || !hasMemoryType )
-          throw new BadRequestError('Do not have the right');
+        for (const type of game_types){
+          const game_type = await this.hr_game_typeRepository.findByCondition({where:{hr_id: hr.id, game_type_id: type}})
+          if (!game_type){
+            isSuccess = 0
+            throw new BadRequestError('Do not have the right to create game id: ' + type);
+          }
+        }
         //create new assessment
-        await this.assessmentRepository.createbyName(hr.id, name,  start_date, end_date)
-        const assessment =  await this.assessmentRepository.findByCondition({where:{hr_id: hr.id, name: name}})
-        const link = 'Padtalent/invite/' + assessment.id
+        const assessment = await this.assessmentRepository.createbyName(hr.id , name,  start_date, end_date)
+        const link = 'Padtalent/test/' + assessment.id
         await this.assessmentRepository.update({link: link},{where:{id: assessment.id}})
-        //insert new assessment's game types 
-        if(logical)
-        await this.assessment_game_typeRepository.create({assessment_id: assessment.id, game_type_id: 1})
-        if(memory)
-        await this.assessment_game_typeRepository.create({assessment_id: assessment.id, game_type_id: 2})
-        return this.setData( 
-            assessment
-          )
-            .setMessage('Success')
-            .responseSuccess(res);
+        //insert new assessment's game types
+        for (const type of game_types){
+          await this.assessment_game_typeRepository.create({assessment_id: assessment.id, game_type_id: type})
+        }
+        if(isSuccess)
+          return this.setData( 
+              {
+                assessment:assessment,
+                game_types: game_types
+              }
+            )
+              .setMessage('Success')
+              .responseSuccess(res);
     } catch (error) {
-      return this.setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
     }
   }
 
@@ -73,30 +77,40 @@ class AssessmentController extends BaseController {
   @Put('/update/:id')
   async update(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
     try {
+        let isSuccess = 1
         const assessDto: AsssessmentDto = req.body;
         const hr = req.hr
-        const {name, logical, memory, start_date, end_date} = assessDto;
-        //res.json({hr_id: hr.id, name: assessDto.name})
-      
+        const {name, game_types, start_date, end_date} = assessDto;
+  
+        //check if hr hr has right to approach the game type
+        for (const type of game_types){
+          const game_type = await this.hr_game_typeRepository.findByCondition({where:{hr_id: hr.id, game_type_id: type}})
+          if (!game_type){
+            isSuccess = 0
+            throw new BadRequestError(' have not the right to create game id ' + type);
+          }
+        }
         const updateAss = await this.assessmentRepository.update(
           {name: name,  start_date: start_date, end_date: end_date},
           {where:{
             id: req.params.id,
           }})
-
-        if(logical)
-        await this.assessment_game_typeRepository.findOrCreateByCondition({where:{assessment_id: req.params.id, game_type_id: 1}})
-        if(memory)
-        await this.assessment_game_typeRepository.findOrCreateByCondition({where:{assessment_id: req.params.id, game_type_id: 2}})
-  
-
-        return this.setData(
-          updateAss
-          )
-            .setMessage('Success')
-            .responseSuccess(res);
+        if(updateAss == 0){
+          isSuccess = 0
+          throw new BadRequestError('can find this assessment');
+        }
+        //update game type
+        for (const type of game_types){
+          await this.assessment_game_typeRepository.findOrCreateByCondition({where:{assessment_id: req.params.id, game_type_id: type}})
+        }
+        if(isSuccess)
+          return this.setData(
+            "update successfully"
+            )
+              .setMessage('Success')
+              .responseSuccess(res);
     } catch (error) {
-      return this.setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
     }
   }
 
@@ -105,39 +119,52 @@ class AssessmentController extends BaseController {
   @Delete('/delete/:id')
   async delete(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
     try {
-        await this.assessmentRepository.deleteById(req.params.id)
-  
-        return this.setData(
-            'Delete assessment successfully'
-          )
-            .setMessage('Success')
-            .responseSuccess(res);
+        const assessment = await this.assessmentRepository.findById(req.params.id) 
+        if(assessment){
+          await this.assessmentRepository.deleteById(req.params.id)
+          return this.setData(
+              'Delete assessment successfully'
+            )
+              .setMessage('Success')
+              .responseSuccess(res);
+        }
+        else{
+          throw new BadRequestError('not found')
+        }
     } catch (error) {
-      return this.setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
     }
   }
 
-  
+  @Authorized()
+  @UseBefore(AuthMiddleware)
+  @Post('/restore/:id')
+  async restore(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
+    try {
+        await this.assessmentRepository.restore({where:{id: req.params.id}})
+        return this.setData(
+            'restore assessment successfully'
+          )
+            .setMessage('Success')
+            .responseSuccess(res);
+        
+    } catch (error) {
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
+    }
+  }
+
   @Authorized()
   @UseBefore(AuthMiddleware)
   @Get('/result/:id')
   async getResult(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
     try {
+        const transporter = this
         const id = req.params.id
         const {count, rows} = await this.testRepository.getByCondition({assessment_id: id}, 0, 100)
        
-        // const result = rows.map(function(test){
-          
-        //   return ({
-        //   id: test.id,
-        //   candidate_id: test.candidate_id,
-        //   game_type_id: test.game_type_id,
-        //   result: test.result})
-        // })
-
         const results = await Promise.all(rows.map(async function(test){
           const candidate_id = test.candidate_id
-          const candidate = await Candidate.findByPk(candidate_id)
+          const candidate = await transporter.candidateRepository.findById(candidate_id)
           const candidate_email = candidate.email
           console.log(candidate.email)
           return ({
@@ -148,8 +175,6 @@ class AssessmentController extends BaseController {
             result: test.result
           })
           }))
-
-
           return this.setData(
           results
           )
@@ -184,6 +209,101 @@ class AssessmentController extends BaseController {
       return this.setCode(error?.status || 500)
         .setMessage(error?.message || 'Internal server error')
         .responseErrors(res);
+    }
+  }
+
+  @Authorized()
+  @UseBefore(AuthMiddleware)
+  @Get('/:id/link')
+  async getLink(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
+    try {
+        let isSuccess = 1
+        const hr = req.hr
+        const assessment = await this.assessmentRepository.findById(req.params.id)
+        if(assessment.hr_id != hr.id){
+          isSuccess = 0
+          throw new BadRequestError('this hr don t have the right to access this ruote');
+        }
+        if (assessment){
+        if (isSuccess)
+          return this.setData({
+              link: assessment.link
+            })
+              .setCode(200)
+              .setMessage('Success')
+              .responseSuccess(res);
+        } else {
+          throw new BadRequestError('Error Assessment');
+        }
+    } catch (error) {
+      return this.setCode(error?.status || 500)
+        .setMessage(error?.message || 'Internal server error')
+        .responseErrors(res);
+    }
+  }
+
+  @Authorized()
+  @UseBefore(AuthMiddleware)
+  @Post('/:id/send-email')
+  async inviteByEmail(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
+    try {
+      const email = req.body.email
+
+      await this.candidateRepository.create({email:email})
+      const assessment = await this.assessmentRepository.findByCondition({id: req.params.id})
+
+      var transporter =  nodemailer.createTransport({ // config mail server
+        host: 'smtp.gmail.com',
+        port: 465,
+        secure: true,
+        auth: {
+            user: 'thideptrai123tq@gmail.com',
+            pass: 'asmfjykwiexxckkg'
+        }
+      });
+      var mainOptions = { // thiết lập đối tượng, nội dung gửi mail
+      from: 'Paditech',
+      to: email,
+      subject: 'Test inviatation',
+      text: 'You have got a invite to test' ,
+      html: '<p>link: <a>' + assessment.link +'</a></p>'
+    }
+
+    transporter.sendMail(mainOptions, function(err, info){
+      if (err) {
+          console.log(err);
+      } else {
+          console.log('Message sent: ' +  info.response);
+      }
+  });
+      
+      return this.setData( 
+            "invite successfully"
+          )
+            .setMessage('Success')
+            .responseSuccess(res);
+    } catch (error) {
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
+    }
+  }
+
+  @Authorized()
+  @UseBefore(AuthMiddleware)
+  @Post('/:id/upload-list')
+  async uploadCandidateList(@Req() req: AuthRequest, @Res() res: Response, next: NextFunction) {
+    try {
+      const data = req.body
+      const list = await this.candidateRepository.bulkCreate(data)
+      for(const candidate of list){
+        await this.candidates_assessmentsRepository.create({candidate_id: candidate.id, assessment_id: req.params.id})
+      }
+      return this.setData( 
+            list
+          )
+            .setMessage('Success')
+            .responseSuccess(res);
+    } catch (error) {
+      return this.setCode(error?.status || 500).setStack(error.stack).setMessage(error?.message || 'Internal server error').responseErrors(res);
     }
   }
 }
